pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pupa-counter"
version = "0.1.0"
description = "CPU-first pupa counting pipeline for scanned images."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "numpy>=1.24",
  "opencv-python-headless>=4.9",
  "scipy>=1.10",
  "scikit-image>=0.21",
  "scikit-learn>=1.3",
  "pandas>=2.0",
  "openpyxl>=3.1",
  "Pillow>=10.0",
  "PyYAML>=6.0",
  "PyMuPDF>=1.23",
  "cellpose>=3.1,<4",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.scripts]
pupa-counter = "pupa_counter.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
